memecah kode menjadi lebih atomic (modular) agar gampang di maintain
