+++ src/components/users/UserModal.jsx
@@ -1,0 +1,136 @@
+import { FaEdit, FaTrashAlt } from "react-icons/fa";
+
+export const UserModal = ({ isOpen, onClose, selectedUser, handleSubmit }) => {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 flex items-center justify-center z-50 bg-black bg-opacity-50 p-4">
+      <div className="bg-white rounded-lg p-6 w-full max-w-md max-h-[90vh] overflow-y-auto">
+        <h3 className="text-xl font-semibold mb-4">
+          {selectedUser ? 'Edit User' : 'Add New User'}
+        </h3>
+        <form className="space-y-4">
+          {/* Nama */}
+          <div>
+            <label htmlFor="nama" className="block text-sm font-medium text-gray-700 mb-1">
+              Full Name
+            </label>
+            <input
+              type="text"
+              id="nama"
+              name="nama"
+              defaultValue={selectedUser?.nama || ''}
+              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
+              placeholder="Enter full name"
+            />
+          </div>
+          {/* Email */}
+          <div>
+            <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
+              Email Address
+            </label>
+            <input
+              type="email"
+              id="email"
+              name="email"
+              defaultValue={selectedUser?.email || ''}
+              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
+              placeholder="Enter email address"
+            />
+          </div>
+          {/* Role */}
+          <div>
+            <label htmlFor="role" className="block text-sm font-medium text-gray-700 mb-1">
+              Role
+            </label>
+            <select
+              id="role"
+              name="role"
+              defaultValue={selectedUser?.role || ''}
+              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
+            >
+              <option value="">Select role</option>
+              <option value="Admin">Admin</option>
+              <option value="User">User</option>
+              <option value="Manager">Manager</option>
+              <option value="Operator">Operator</option>
+            </select>
+          </div>
+          {/* Status */}
+          <div>
+            <label htmlFor="status" className="block text-sm font-medium text-gray-700 mb-1">
+              Status
+            </label>
+            <select
+              id="status"
+              name="status"
+              defaultValue={selectedUser?.status || ''}
+              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
+            >
+              <option value="">Select status</option>
+              <option value="Active">Active</option>
+              <option value="Inactive">Inactive</option>
+              <option value="Pending">Pending</option>
+            </select>
+          </div>
+          {/* Jenis Kelamin */}
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Gender
+            </label>
+            <div className="flex gap-4">
+              <label className="inline-flex items-center">
+                <input
+                  type="radio"
+                  name="jenis_kelamin"
+                  value="Laki-laki"
+                  defaultChecked={selectedUser?.jenis_kelamin === 'Laki-laki'}
+                  className="form-radio h-4 w-4 text-indigo-600"
+                />
+                <span className="ml-2">Laki-laki</span>
+              </label>
+              <label className="inline-flex items-center">
+                <input
+                  type="radio"
+                  name="jenis_kelamin"
+                  value="Perempuan"
+                  defaultChecked={selectedUser?.jenis_kelamin === 'Perempuan'}
+                  className="form-radio h-4 w-4 text-indigo-600"
+                />
+                <span className="ml-2">Perempuan</span>
+              </label>
+            </div>
+          </div>
+          {/* Tanggal Lahir */}
+          <div>
+            <label htmlFor="tanggal_lahir" className="block text-sm font-medium text-gray-700 mb-1">
+              Birth Date
+            </label>
+            <input
+              type="date"
+              id="tanggal_lahir"
+              name="tanggal_lahir"
+              defaultValue={selectedUser?.tanggal_lahir || ''}
+              className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
+            />
+          </div>
+          <div className="flex justify-end gap-3 mt-6">
+            <button
+              type="button"
+              onClick={onClose}
+              className="px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50"
+            >
+              Cancel
+            </button>
+            <button
+              type="submit"
+              className="px-4 py-2 bg-indigo-600 text-white rounded-md hover:bg-indigo-700"
+            >
+              {selectedUser ? 'Update User' : 'Add User'}
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+};
